# server/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sheeps API",
    "version": "1.0.0",
    "description": "Sheeps 手游后端 API 文档。\n\n**约定**：除明确标注的 GET 接口外，大部分接口的请求体为「加密信封」，结构形如 `{ \"encrypted\": \"<base64>\" }`，由服务端 `decryptRequest` 中间件解密为真实 JSON 后再分发。Swagger 中 requestBody 给出的均为「解密后的真实 JSON 示例」。\n\n**鉴权**：带 `/api/admin/*`、`/api/match/*` 等接口的访问令牌通过 `Authorization: Bearer <token>` 头传递（具体见各接口说明）。"
  },
...
  "paths": {
...
    "/api/admin/app-versions": {
      "get": {
        "tags": ["admin"],
        "summary": "应用版本列表（后台）",
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "example": { "success": true, "list": [] } } } }
        }
      },
      "post": {
        "tags": ["admin"],
        "summary": "新增/更新应用版本（status=1 即发布）",
        "requestBody": {
          "content": {
            "application/json": {
              "example": { "version_code": 12, "version_name": "1.1.2", "download_url": "https://...", "status": 1, "force_update": 0 }
            }
          }
        },
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "example": { "success": true } } } }
        }
      }
    },
    "/api/admin/app-versions/{code}": {
      "delete": {
        "tags": ["admin"],
        "summary": "撤回指定版本",
        "parameters": [
          { "name": "code", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "example": { "success": true } } } }
        }
      }
    },
    "/api/admin/notices": {
      "get": {
        "tags": ["admin"],
        "summary": "公告列表（后台）",
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "example": { "success": true, "list": [] } } } }
        }
      },
      "post": {
        "tags": ["admin"],
        "summary": "新建公告",
        "requestBody": {
          "content": { "application/json": { "example": { "title": "维护通知", "content": "...", "platform": "all" } } }
        },
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "example": { "success": true } } } }
        }
      },
      "delete": {
        "tags": ["admin"],
        "summary": "删除公告（路径带 ?id=）",
        "parameters": [
          { "name": "id", "in": "query", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "example": { "success": true } } } }
        }
      }
    }
  }
}